reformated Sharvand map

+++ app/src/main/java/ir/parsiot/pokdis/Localization/ParticleFilter/ParticleFilterMath.java
@@ -1,7 +1,7 @@
 package ir.parsiot.pokdis.Localization.ParticleFilter;
 
 public class ParticleFilterMath {
-    public static double distance(float x1, float y1, float x2, float y2) {
+    public static double distance(double x1, double y1, double x2, double y2) {
         return Math.sqrt((x1 - x2) * (x1 - x2) + (y1 - y2) * (y1 - y2));
     }
